pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-free-swap"
version = "0.1.0"
description = "OpenAI-compatible proxy that routes through free-tier AI providers with priority-based fallback"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sse-starlette>=1.8",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "langchain-core>=0.3",
    "langchain-google-genai>=2.0",
    "langchain-openai>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-qwq>=0.1",
    "langchain-openrouter>=0.1",
]

[project.scripts]
ai-free-swap = "ai_free_swap.main:main"

[tool.setuptools.packages.find]
where = ["src"]
